+++ CHANGELOG.md
@@ -18,6 +18,7 @@ In place of release version numbers, we organize via deploys to Production (by D
   - Index views: indicate resident_status & x of y models
     - Using new residencies/_property & _resident icon list views
   - Property#show: lists Lot_fee_paid? (as well as date) 
+  - Lot#paid_on formatted time_ago_in_words
 - Seed
   - Rename JimR to "Peter PartialPayment"
   - Add (random) comments